Copy L2_voronoi_traits_2.h to L2_segment_voronoi_traits_2.h, make fn 10 copy of fn 5
To make FSVD using envelopes like it's done in fn 5 for points, it's necessary to create a new traits file. For now it was just copied from the one used for points.
Fn 10 (the one that should do FSVD) just does the same as fn 5 (FDV for points) for now.
It compiled and works in Ipe.