Change Rat/Alg logic, change arc to segments logic
Square root did not work, it crashed because the number types did not support it. It compiles, but it does not work on Ipe; it should draw a circular arc defined by three points of the two segments (two endpoints of one and an endpoint of the other) just for testing. It appears as if the arc to segments works, but no segments are to be found in the list when drawing.