Start computing unbounded edges of bisector
Use convex hull to find exterior points, pairs of points not of the same segment are the one defining bisectors that contain the unbounded rays of the plane bisector for two segments.
Convex hull algorithm fails, no idea why for now. (It's a segmentation fault. Am I giving an invalid iterator perhaps?)
For now the Ipelet is changed so that it compiles and draws an edge between the two sources of two segments.