Change function to find ray start point to find just the ray
This way we have all the information about the directions. When continuing the construction of the bisector, we know approximately in which direction it must go: the opposite of the ray from which we start.
The ray is still converted to a very long segment to be added to the list of pieces of the bisector.