Implement function to adjust one endpoint of a segment (controlled)
Two flags determine in which direction to approximate the endpoint (towards the positive or negative side of the supporting line of the segment) and which one of the two segments needs to be approximated. The other endpoint is already determined to be actually a rational point by construction, and is passed as an argument too.
The function returns a rational segment.